Refactor: HomeView 분리 (#12)
